Tests/LibWeb/Text/expected/UIEvents/KeyEvent-tab.txt

input1 keydown Tab
<INPUT id="input2" >
input2 keydown Tab
<INPUT id="input1" >
